requirements.txt
# Probable.news — Backend Dependencies
# Python 3.11+

# Web
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Database
sqlalchemy>=2.0.25
asyncpg>=0.29.0
psycopg2-binary>=2.9.9
alembic>=1.13.0

# Redis & Workers
redis>=5.0.0
arq>=0.25.0
httpx>=0.26.0

# RSS & Article Extraction
feedparser>=6.0.10
readability-lxml>=0.8.1
lxml>=5.1.0
beautifulsoup4>=4.12.0

# LLM
openai>=1.12.0

# Auth
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0

# Config & Utilities
pydantic[email]>=2.6.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0

# Storage (S3-compatible)
boto3>=1.34.0

# Charts & Data
pandas>=2.2.0
numpy>=1.26.0
plotly>=5.18.0
